Agregar módulo de Reportes y refactorizar sembradoras
- Se introduce reportController y el modelo de informes para la gestión de informes.
- Se añade la migración de informes y se elimina la migración/modelo de sesiones de usuario.
- Se refactorizan los sembradores a archivos modulares para sucursales, clientes, productos, inventario, pagos, ventas, artículos en oferta, usuarios e informes.
- Se actualiza saleController para crear registros de pago automáticamente al crear una venta.
- Se añade compatibilidad con borrado temporal (paranoico) a los modelos de pago y devolución.
- Se limpia y optimiza la lógica de los sembradores en index.js.
