+++ api/src/infrastructure/database/models/Payment.js
@@ -40,6 +40,10 @@ module.exports = (sequelize, DataTypes) => {
         {
             tableName: 'payments',
             timestamps: true,
+            paranoid: true,
+            createdAt: 'created_at',
+            updatedAt: 'updated_at',
+            deletedAt: 'deleted_at',
         }
     );
