+++ api/src/infrastructure/database/models/Report.js
@@ -1,0 +1,52 @@
+module.exports = (sequelize, DataTypes) => {
+  const Report = sequelize.define('Report', {
+    id: {
+      type: DataTypes.UUID,
+      defaultValue: DataTypes.UUIDV4,
+      primaryKey: true,
+      allowNull: false
+    },
+    branch_id: {
+      type: DataTypes.UUID,
+      allowNull: true,
+      references: {
+        model: 'branches',
+        key: 'id'
+      }
+    },
+    type: {
+      type: DataTypes.ENUM('sales', 'inventory', 'performance'),
+      allowNull: false
+    },
+    title: {
+      type: DataTypes.STRING(100),
+      allowNull: false
+    },
+    description: {
+      type: DataTypes.TEXT,
+      allowNull: true
+    },
+    data: {
+      type: DataTypes.JSON,
+      allowNull: false
+    },
+    generated_at: {
+      type: DataTypes.DATE,
+      allowNull: false,
+      defaultValue: DataTypes.NOW
+    }
+  }, {
+    tableName: 'reports',
+    timestamps: true,
+    paranoid: true,
+    createdAt: 'created_at',
+    updatedAt: 'updated_at',
+    deletedAt: 'deleted_at'
+  });
+
+  Report.associate = models => {
+    Report.belongsTo(models.Branch, { foreignKey: 'branch_id', as: 'branch' });
+  };
+
+  return Report;
+};
