+++ api/src/infrastructure/database/models/Return.js
@@ -46,6 +46,10 @@ module.exports = (sequelize, DataTypes) => {
         {
             tableName: 'returns',
             timestamps: true,
+            paranoid: true,
+            createdAt: 'created_at',
+            updatedAt: 'updated_at',
+            deletedAt: 'deleted_at',
         }
     );
