+++ api/src/infrastructure/database/seeders/seedProducts.js
@@ -1,0 +1,173 @@
+const db = require('../models');
+const { Product } = db;
+
+async function seedProducts() {
+  return await Product.bulkCreate([
+   {
+            name: 'PlayStation 5',
+            description: 'Consola de nueva generación con SSD ultra rápido y gráficos 4K',
+            sku: 'PS5-CONSOLE-001',
+            barcode: '711719541028',
+            category: 'Consola',
+            subcategory: 'PlayStation',
+            cost_price: 450.00,
+            unit_price: 599.99,
+            unit_measure: 'unit',
+            weight: 4.5,
+            dimensions: JSON.stringify({ length: 39, width: 26, height: 10.4 }),
+            tax_rate: 0.16,
+            is_active: true,
+            metadata: JSON.stringify({ platform: 'PlayStation', generation: 5, storage: '825GB SSD' }),
+            tags: JSON.stringify(['PS5', 'Sony', 'Nueva Generación', '4K'])
+        },
+        {
+            name: 'Xbox Series X',
+            description: 'La consola Xbox más potente de todos los tiempos con 12 TFLOPS',
+            sku: 'XBOX-SX-001',
+            barcode: '889842640649',
+            category: 'Consola',
+            subcategory: 'Xbox',
+            cost_price: 450.00,
+            unit_price: 599.99,
+            unit_measure: 'unit',
+            weight: 4.45,
+            dimensions: JSON.stringify({ length: 30.1, width: 15.1, height: 15.1 }),
+            tax_rate: 0.16,
+            is_active: true,
+            metadata: JSON.stringify({ platform: 'Xbox', generation: 'Series X', storage: '1TB SSD' }),
+            tags: JSON.stringify(['Xbox', 'Microsoft', 'Series X', '12 TFLOPS'])
+        },
+        {
+            name: 'Nintendo Switch OLED',
+            description: 'Nintendo Switch con pantalla OLED de 7 pulgadas',
+            sku: 'NSW-OLED-001',
+            barcode: '045496882068',
+            category: 'Consola',
+            subcategory: 'Nintendo',
+            cost_price: 280.00,
+            unit_price: 349.99,
+            unit_measure: 'unit',
+            weight: 0.42,
+            dimensions: JSON.stringify({ length: 24.2, width: 10.4, height: 1.4 }),
+            tax_rate: 0.16,
+            is_active: true,
+            metadata: JSON.stringify({ platform: 'Nintendo', type: 'Hybrid', storage: '64GB' }),
+            tags: JSON.stringify(['Nintendo', 'Switch', 'OLED', 'Portátil'])
+        },
+        {
+            name: 'The Legend of Zelda: Tears of the Kingdom',
+            description: 'La épica continuación de Breath of the Wild',
+            sku: 'ZELDA-TOTK-NSW',
+            barcode: '045496596899',
+            category: 'Videojuego',
+            subcategory: 'Aventura',
+            cost_price: 42.00,
+            unit_price: 69.99,
+            unit_measure: 'unit',
+            weight: 0.1,
+            tax_rate: 0.16,
+            is_active: true,
+            metadata: JSON.stringify({ platform: 'Nintendo Switch', genre: 'Aventura', rating: 'E10+' }),
+            tags: JSON.stringify(['Zelda', 'Nintendo', 'Aventura', 'Open World'])
+        },
+        {
+            name: 'God of War Ragnarök',
+            description: 'La conclusión de la saga nórdica de Kratos y Atreus',
+            sku: 'GOW-RAGNAROK-PS5',
+            barcode: '711719543294',
+            category: 'Videojuego',
+            subcategory: 'Acción',
+            cost_price: 42.00,
+            unit_price: 69.99,
+            unit_measure: 'unit',
+            weight: 0.1,
+            tax_rate: 0.16,
+            is_active: true,
+            metadata: JSON.stringify({ platform: 'PlayStation 5', genre: 'Acción', rating: 'M' }),
+            tags: JSON.stringify(['God of War', 'PlayStation', 'Acción', 'Aventura'])
+        },
+        {
+            name: 'Halo Infinite',
+            description: 'El regreso del Master Chief en una nueva aventura',
+            sku: 'HALO-INF-XBOX',
+            barcode: '889842640632',
+            category: 'Videojuego',
+            subcategory: 'Shooter',
+            cost_price: 35.00,
+            unit_price: 59.99,
+            unit_measure: 'unit',
+            weight: 0.1,
+            tax_rate: 0.16,
+            is_active: true,
+            metadata: JSON.stringify({ platform: 'Xbox Series X/S', genre: 'FPS', rating: 'T' }),
+            tags: JSON.stringify(['Halo', 'Xbox', 'FPS', 'Multijugador'])
+        },
+        {
+            name: 'DualSense Wireless Controller',
+            description: 'Control inalámbrico para PS5 con retroalimentación háptica',
+            sku: 'DS-CTRL-WHITE',
+            barcode: '711719541172',
+            category: 'Accesorio',
+            subcategory: 'Control',
+            cost_price: 45.00,
+            unit_price: 69.99,
+            unit_measure: 'unit',
+            weight: 0.28,
+            tax_rate: 0.16,
+            is_active: true,
+            metadata: JSON.stringify({ platform: 'PlayStation 5', type: 'Controller', wireless: true }),
+            tags: JSON.stringify(['DualSense', 'PlayStation', 'Control', 'Háptico'])
+        },
+        {
+            name: 'Xbox Wireless Controller',
+            description: 'Control inalámbrico para Xbox Series X/S',
+            sku: 'XBOX-CTRL-BLACK',
+            barcode: '889842640656',
+            category: 'Accesorio',
+            subcategory: 'Control',
+            cost_price: 40.00,
+            unit_price: 59.99,
+            unit_measure: 'unit',
+            weight: 0.29,
+            tax_rate: 0.16,
+            is_active: true,
+            metadata: JSON.stringify({ platform: 'Xbox Series X/S', type: 'Controller', wireless: true }),
+            tags: JSON.stringify(['Xbox', 'Control', 'Inalámbrico'])
+        },
+        {
+            name: 'Steam Deck 64GB',
+            description: 'Consola portátil para juegos de Steam',
+            sku: 'STEAM-DECK-64',
+            barcode: '814585021547',
+            category: 'PC Gaming',
+            subcategory: 'Portátil',
+            cost_price: 320.00,
+            unit_price: 399.99,
+            unit_measure: 'unit',
+            weight: 0.67,
+            dimensions: JSON.stringify({ length: 29.8, width: 11.7, height: 4.9 }),
+            tax_rate: 0.16,
+            is_active: true,
+            metadata: JSON.stringify({ platform: 'PC', type: 'Handheld', storage: '64GB eMMC' }),
+            tags: JSON.stringify(['Steam', 'Valve', 'PC Gaming', 'Portátil'])
+        },
+        {
+            name: 'PlayStation Store $50',
+            description: 'Tarjeta de regalo para PlayStation Store',
+            sku: 'PSN-CARD-50',
+            barcode: '711719513637',
+            category: 'Tarjeta de regalo',
+            subcategory: 'PlayStation',
+            cost_price: 45.00,
+            unit_price: 50.00,
+            unit_measure: 'unit',
+            weight: 0.01,
+            tax_rate: 0.00,
+            is_active: true,
+            metadata: JSON.stringify({ platform: 'PlayStation', value: 50, currency: 'USD' }),
+            tags: JSON.stringify(['PlayStation', 'PSN', 'Tarjeta Regalo'])
+        }
+  ]);
+}
+
+module.exports = seedProducts;
